requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
cloudinary==1.36.0
Django==4.2.6
django-cloudinary-storage==0.3.0
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-htmx==1.17.0
django-phonenumber-field==7.2.0
idna==3.4
phonenumberslite==8.13.23
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
